pyproject.toml
[project]
name = "learn-pydantic-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Lukas Hamm",email = "lukas@lukashamm.dev"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai-slim[mistral] (>=1.0.17,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "LearnPydanticAI", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
